+++ src/lint/index.ts
@@ -1,1 +1,6 @@
 export { Lint } from "./src/index.js";
+export {
+    forEachScientificNotationToken,
+    toPlainDecimalFromScientificLiteral,
+    trimInsignificantFractionalZeros
+} from "./src/index.js";
