+++ src/lint/src/malformed/index.ts
@@ -1,7 +1,9 @@
 export {
     forEachScientificNotationToken,
     isScientificNotationBoundary,
-    SCIENTIFIC_NOTATION_PATTERN
+    SCIENTIFIC_NOTATION_PATTERN,
+    toPlainDecimalFromScientificLiteral,
+    trimInsignificantFractionalZeros
 } from "./scientific-notation-scan.js";
 export type { CommentFixResult } from "./source-preprocessing.js";
 export { fixMalformedComments, recoverParseSourceFromMissingBrace } from "./source-preprocessing.js";
